refactor: Extract reps from HTTP handlers.
Finish moving response representations out of the HTTP handlers module
and into the `reps` module.
