+++ src/ledger/http/mod.rs
@@ -1,8 +1,6 @@
-use std::{collections::HashSet, iter::FromIterator};
+use std::iter::FromIterator;
 
-use chrono::{DateTime, NaiveDate, Utc};
 use rocket::{http::Status, serde::json::Json, Route};
-use serde::{Deserialize, Serialize};
 use tracing::error;
 use uuid::Uuid;
 
@@ -14,7 +12,7 @@ use crate::{
 
 use super::{
     commands::{postgres::PostgresCommands, TransactionCommands, UpdateTransactionError},
-    domain::{self, transactions::NewTransactionError},
+    domain,
     queries::{postgres::PostgresQueries, CurrencyQueries, TransactionQueries},
 };
 
@@ -28,74 +26,6 @@ pub fn routes() -> Vec<Route> {
         get_transactions,
         update_transaction,
     ]
-}
-
-#[derive(Deserialize)]
-pub struct NewTransaction {
-    pub date: chrono::NaiveDate,
-    pub payee: String,
-    pub notes: Option<String>,
-    pub entries: Vec<NewTransactionEntry>,
-}
-
-impl NewTransaction {
-    pub fn used_currency_codes(&self) -> HashSet<String> {
-        self.entries
-            .iter()
-            .filter_map(|entry| {
-                entry
-                    .amount
-                    .as_ref()
-                    .map(|amount| amount.currency.to_owned())
-            })
-            .collect()
-    }
-}
-
-#[derive(Deserialize)]
-pub struct NewTransactionEntry {
-    pub account: String,
-    pub amount: Option<reps::CurrencyAmount>,
-}
-
-#[derive(Serialize)]
-pub struct Transaction {
-    pub id: Uuid,
-    pub date: NaiveDate,
-    pub payee: String,
-    pub notes: String,
-    pub entries: Vec<TransactionEntry>,
-    pub created_at: DateTime<Utc>,
-    pub updated_at: DateTime<Utc>,
-}
-
-impl From<&domain::transactions::Transaction> for Transaction {
-    fn from(domain: &domain::transactions::Transaction) -> Self {
-        Self {
-            id: domain.id,
-            date: domain.date,
-            payee: domain.payee.to_owned(),
-            notes: domain.notes.to_owned(),
-            entries: domain.entries.iter().map(|entry| entry.into()).collect(),
-            created_at: domain.created_at,
-            updated_at: domain.updated_at,
-        }
-    }
-}
-
-#[derive(Serialize)]
-pub struct TransactionEntry {
-    pub account: String,
-    pub amount: reps::CurrencyAmount,
-}
-
-impl From<&domain::transactions::TransactionEntry> for TransactionEntry {
-    fn from(domain: &domain::transactions::TransactionEntry) -> Self {
-        Self {
-            account: domain.account().to_string(),
-            amount: domain.amount().into(),
-        }
-    }
 }
 
 #[delete("/transactions/<transaction_id>")]
@@ -121,7 +51,7 @@ async fn delete_transaction(
 
 #[derive(Responder)]
 pub enum GetTransactionResponse {
-    Ok(Json<Transaction>),
+    Ok(Json<reps::Transaction>),
     #[response(status = 404)]
     NotFound(Json<ErrorRep>),
 }
@@ -162,7 +92,7 @@ async fn get_transaction(
 async fn get_transactions(
     session: Session,
     db: PostgresConn,
-) -> Result<Json<reps::ResourceCollection<Transaction>>, ApiError> {
+) -> Result<Json<reps::ResourceCollection<reps::Transaction>>, ApiError> {
     let queries = PostgresQueries(&db);
 
     match queries.latest_transactions(session.user_id()).await {
@@ -183,7 +113,7 @@ async fn get_transactions(
 #[derive(Responder)]
 pub enum CreateTransactionResponse {
     #[response(status = 201)]
-    Created(Json<Transaction>),
+    Created(Json<reps::Transaction>),
     #[response(status = 400)]
     BadRequest(Json<reps::TransactionValidationError>),
 }
@@ -203,7 +133,7 @@ impl From<reps::TransactionValidationError> for CreateTransactionResponse {
 #[post("/transactions", data = "<new_transaction>")]
 async fn create_transaction(
     session: Session,
-    new_transaction: Json<NewTransaction>,
+    new_transaction: Json<reps::NewTransaction>,
     db: PostgresConn,
 ) -> Result<CreateTransactionResponse, ApiError> {
     let queries = PostgresQueries(&db);
@@ -218,56 +148,9 @@ async fn create_transaction(
         }
     };
 
-    let mut parsed_entries = Vec::with_capacity(new_transaction.entries.len());
-    for new_entry in new_transaction.entries.iter() {
-        let parsed_amount = match &new_entry.amount {
-            None => None,
-            Some(amount_rep) => {
-                if let Some(currency) = used_currencies.get(&amount_rep.currency) {
-                    let parse_result = domain::currency::CurrencyAmount::from_str(
-                        currency.clone(),
-                        &amount_rep.value,
-                    );
-
-                    match parse_result {
-                        Ok(amount) => Some(amount),
-                        Err(error) => {
-                            return Ok(reps::TransactionValidationError::from(error).into())
-                        }
-                    }
-                } else {
-                    return Ok(reps::TransactionValidationError {
-                        message: Some(format!(
-                            "The currency code '{}' is unrecognized.",
-                            &amount_rep.currency
-                        )),
-                    }
-                    .into());
-                }
-            }
-        };
-
-        parsed_entries.push(domain::transactions::NewTransactionEntry {
-            account: new_entry.account.clone(),
-            amount: parsed_amount,
-        });
-    }
-
-    let transaction = match domain::transactions::NewTransaction::new(
-        session.user_id(),
-        new_transaction.date,
-        new_transaction.payee.clone(),
-        new_transaction.notes.clone(),
-        parsed_entries,
-    ) {
+    let transaction = match new_transaction.try_into_domain(session.user_id(), used_currencies) {
         Ok(t) => t,
-        Err(NewTransactionError::Unbalanced(_)) => {
-            return Ok(CreateTransactionResponse::BadRequest(Json(
-                reps::TransactionValidationError {
-                    message: Some("The entries in the transaction are unbalanced.".to_owned()),
-                },
-            )))
-        }
+        Err(error) => return Ok(error.into()),
     };
 
     let ledger_commands = PostgresCommands(&db);
@@ -287,7 +170,7 @@ async fn create_transaction(
 #[derive(Responder)]
 pub enum UpdateTransactionResponse {
     #[response(status = 200)]
-    Updated(Json<Transaction>),
+    Updated(Json<reps::Transaction>),
     #[response(status = 400)]
     BadRequest(Json<reps::TransactionValidationError>),
     #[response(status = 404)]
@@ -310,7 +193,7 @@ impl From<reps::TransactionValidationError> for UpdateTransactionResponse {
 async fn update_transaction(
     session: Session,
     transaction_id: Uuid,
-    updated_transaction: Json<NewTransaction>,
+    updated_transaction: Json<reps::NewTransaction>,
     db: PostgresConn,
 ) -> Result<UpdateTransactionResponse, ApiError> {
     let queries = PostgresQueries(&db);
@@ -325,55 +208,10 @@ async fn update_transaction(
         }
     };
 
-    let mut parsed_entries = Vec::with_capacity(updated_transaction.entries.len());
-    for new_entry in updated_transaction.entries.iter() {
-        let parsed_amount = match &new_entry.amount {
-            None => None,
-            Some(amount_rep) => {
-                if let Some(currency) = used_currencies.get(&amount_rep.currency) {
-                    let parse_result = domain::currency::CurrencyAmount::from_str(
-                        currency.clone(),
-                        &amount_rep.value,
-                    );
-
-                    match parse_result {
-                        Ok(amount) => Some(amount),
-                        Err(error) => {
-                            return Ok(reps::TransactionValidationError::from(error).into())
-                        }
-                    }
-                } else {
-                    return Ok(reps::TransactionValidationError {
-                        message: Some(format!(
-                            "The currency code '{}' is unrecognized.",
-                            &amount_rep.currency
-                        )),
-                    }
-                    .into());
-                }
-            }
-        };
-
-        parsed_entries.push(domain::transactions::NewTransactionEntry {
-            account: new_entry.account.clone(),
-            amount: parsed_amount,
-        });
-    }
-
-    let transaction = match domain::transactions::NewTransaction::new(
-        session.user_id(),
-        updated_transaction.date,
-        updated_transaction.payee.clone(),
-        updated_transaction.notes.clone(),
-        parsed_entries,
-    ) {
+    let transaction = match updated_transaction.try_into_domain(session.user_id(), used_currencies)
+    {
         Ok(t) => t,
-        Err(NewTransactionError::Unbalanced(_)) => {
-            return Ok(reps::TransactionValidationError {
-                message: Some("The entries in the transaction are unbalanced.".to_string()),
-            }
-            .into())
-        }
+        Err(error) => return Ok(error.into()),
     };
 
     let ledger_commands = PostgresCommands(&db);
